+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Valkey GLIDE C# client will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## 1.2.0
+
+### Added
+
+- Custom socket address resolution support via callback (#392)
+
 ## 1.1.0
 
 ### Added